+++ packages/ra-no-code/src/ResourceConfiguration/index.ts
@@ -2,3 +2,4 @@ export * from './getFieldDefinitionsFromRecords';
 export * from './useResource';
 export * from './useResourceConfigurations';
 export * from './ResourceConfigurationContext';
+export * from './ResourceConfigurationProvider';
