+++ packages/ra-no-code/src/index.ts
@@ -1,4 +1,4 @@
 export * from './Admin';
 export * from './builders';
-export * from './ResourceBuilderContext';
+export * from './ResourceConfiguration';
 export * from './ui';
